refactor: extract bundled extension manifest parser
